+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### main / head
 
+* `<` now dedents current line
+
 ### v0.0.3 | 17/Mar/2026
 * Use `;e` to open a different file from within iris
 * Keyboard commands `tt`, `tn`, `tp` to create and switch between tabs